[refactor] change compare-tabs utils name
- to include other functions, change compare-tabs utils name
